Fix #39: Validation input on the phone input field
